Sound system base

+++ MathHelper.cs
@@ -17,7 +17,7 @@ namespace EngineSimulator {
             return value;
         }
 
-        public static double Clamp(float value, float min, float max) {
+        public static float Clamp(float value, float min, float max) {
             if (value < min) return min;
             if (value > max) return max;
             return value;
